Authorization on deleting listing

+++ App/Controllers/ListingController.php
@@ -4,6 +4,8 @@ namespace App\Controllers;
 
 use Framework\Database;
 use Framework\Validation;
+use Framework\Session;
+use Framework\Authorization;
 
 class ListingController {
 
@@ -19,7 +21,7 @@ class ListingController {
 
     public function index(){
         
-        $listings = $this->db->query('SELECT * FROM listings')->fetchAll();
+        $listings = $this->db->query('SELECT * FROM listings ORDER BY created_at DESC')->fetchAll();
 
         loadView('listings/index', [
             'listings' => $listings
@@ -59,7 +61,7 @@ class ListingController {
 
         $newListingData = array_intersect_key($_POST, array_flip($allowedfields));
 
-        $newListingData['user_id'] = 1;
+        $newListingData['user_id'] = Session::get('user')['id'];
 
         $newListingData = array_map('sanitize', $newListingData);
 
@@ -125,9 +127,16 @@ public function destroy($params) {
 
     $listings = $this->db->query('SELECT * FROM listings WHERE id = :id', $params)->fetch();
 
+    //Check if listing exist
     if (!$listings) {
         ErrorController::notFound('Listing not found');
         return;
+    }
+
+    //Authorization
+    if(!Authorization::isOwner($listings->user_id)){
+        $_SESSION['error_message'] = 'You are not authorized to delete this listing';
+        return redirect('/listings/' . $listings->id);  
     }
 
     $this->db->query('DELETE FROM listings WHERE id = :id', $params);
